update inputs models and class
