+++ libs/src/lib/dom/components/form/models/input-base.types.ts
@@ -1,0 +1,93 @@
+import { MatFormFieldAppearance } from '@angular/material/form-field';
+import { InputType } from './input.types';
+
+export type Appearance = 'none' | MatFormFieldAppearance;
+
+export interface BaseInputField {
+  /**
+   * This prop helps users to fill forms faster, especially on mobile devices.
+   * The name can be confusing, as it's more like an autofill.
+   * You can learn more about it [following the specification](https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#autofill).
+   */
+
+  /**
+   * Name attribute of the `input` element.
+   */
+  name: string;
+
+  appearance?: Appearance;
+
+  autoComplete?: boolean;
+  /**
+   * If `true`, the `input` element is focused during the first mount.
+   * @default false
+   */
+  autoFocus?: boolean;
+
+  /**
+   * Override or extend the styles applied to the component.
+   */
+  classes?: string;
+  /**
+   * The color of the component.
+   * It supports both default and custom theme colors, which can be added as shown in the
+   * [palette customization guide](https://mui.com/material-ui/customization/palette/#custom-colors).
+   * @default 'primary'
+   */
+  color?: 'primary' | 'secondary' | 'error' | 'info' | 'success' | 'warning';
+  /**
+   * The default value.
+   */
+  defaultValue?: unknown;
+  /**
+   * If `true`, the component is disabled.
+   * @default false
+   */
+  disabled?: boolean;
+  /**
+   * If `true`, the label is displayed in an error state.
+   * @default false
+   */
+  error?: boolean;
+  /**
+   * Props applied to the [`FormHelperText`](/material-ui/api/form-helper-text/) element.
+   */
+  // FormHelperTextProps?: Partial<FormHelperTextProps>;
+  /**
+   * If `true`, the input will take up the full width of its container.
+   * @default false
+   */
+  fullWidth?: boolean;
+  /**
+   * The helper text content.
+   */
+  helperText?: string;
+  /**
+   * The id of the `input` element.
+   * Use this prop to make `label` and `helperText` accessible for screen readers.
+   */
+  placeHolder?: string;
+  /**
+   * The label content.
+   */
+  label?: string;
+
+  // onBlur?: InputBaseProps['onBlur'];
+  // onFocus?: StandardInputProps['onFocus'];
+  /**
+   * The short hint displayed in the `input` before the user enters a value.
+   */
+  placeholder?: string;
+  /**
+   * If `true`, the label is displayed as required and the `input` element is required.
+   * @default false
+   */
+  required?: boolean;
+
+  /**
+   * Type of the `input` element. It should be [a valid HTML5 input type](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#Form_%3Cinput%3E_types).
+   */
+  type?: InputType;
+
+  icon?: string;
+}
